+++ dune/common/diagonalmatrix.hh
@@ -20,10 +20,13 @@
 
 #include <dune/common/boundschecking.hh>
 #include <dune/common/densematrix.hh>
+#include <dune/common/dynmatrix.hh>
 #include <dune/common/exceptions.hh>
 #include <dune/common/fmatrix.hh>
+#include <dune/common/ftraits.hh>
 #include <dune/common/fvector.hh>
 #include <dune/common/genericiterator.hh>
+#include <dune/common/matrixconcepts.hh>
 #include <dune/common/typetraits.hh>
 
 
@@ -114,8 +117,7 @@ namespace Dune {
      */
     DiagonalMatrix (std::initializer_list<K> const &l)
     {
-      std::copy_n(l.begin(), std::min(static_cast<std::size_t>(rows),
-                                      l.size()),
+      std::copy_n(l.begin(), std::min<std::size_t>(rows, l.size()),
                  diag_.begin());
     }
 
@@ -474,7 +476,39 @@ namespace Dune {
       return result;
     }
 
+    /**
+     * \brief Multiply a diagonal matrix with a dense matrix.
+     *
+     * The result of this multiplication is either a `FieldMatrix` if the
+     * `matrixB` is a matrix with static size, or a `DynamicMatrix`. This
+     * overload is deactivated for `matrixB` being a `FieldMatrix` since this
+     * is already covered by the corresponding overload of the `operator*` in
+     * the `FieldMatrix` class.
+     */
+    template <class OtherMatrix,
+      std::enable_if_t<(Impl::IsDenseMatrix<OtherMatrix>::value), int> = 0,
+      std::enable_if_t<(not Impl::IsFieldMatrix<OtherMatrix>::value), int> = 0>
+    friend auto operator* ( const DiagonalMatrix& matrixA,
+                            const OtherMatrix& matrixB)
+    {
+      using OtherField = typename FieldTraits<OtherMatrix>::field_type;
+      using F = typename PromotionTraits<field_type, OtherField>::PromotedType;
 
+      auto result = [&]{
+        if constexpr (Impl::IsStaticSizeMatrix_v<OtherMatrix>) {
+          static_assert(n == OtherMatrix::rows);
+          return FieldMatrix<F, n, OtherMatrix::cols>{};
+        } else {
+          assert(n == matrixB.N());
+          return DynamicMatrix<F>{n,matrixB.M()};
+        }
+      }();
+
+      for (int i = 0; i < result.N(); ++i)
+        for (int j = 0; j < result.M(); ++j)
+          result[i][j] = matrixA.diagonal(i) * matrixB[i][j];
+      return result;
+    }
 
     //===== sizes
 
@@ -639,6 +673,31 @@ namespace Dune {
                             const DiagonalMatrix<OtherScalar, 1>& matrixB)
     {
       return DiagonalMatrix<typename PromotionTraits<K,OtherScalar>::PromotedType, 1>{matrixA.diagonal(0)*matrixB.diagonal(0)};
+    }
+
+    template <class OtherMatrix,
+      std::enable_if_t<(Impl::IsDenseMatrix<OtherMatrix>::value), int> = 0,
+      std::enable_if_t<(not Impl::IsFieldMatrix<OtherMatrix>::value), int> = 0>
+    friend auto operator* ( const DiagonalMatrix& matrixA,
+                            const OtherMatrix& matrixB)
+    {
+      using OtherField = typename FieldTraits<OtherMatrix>::field_type;
+      using F = typename PromotionTraits<K, OtherField>::PromotedType;
+
+      auto result = [&]{
+        if constexpr (Impl::IsStaticSizeMatrix_v<OtherMatrix>) {
+          static_assert(1 == OtherMatrix::rows);
+          return FieldMatrix<F, 1, OtherMatrix::cols>{};
+        } else {
+          assert(1 == matrixB.N());
+          return DynamicMatrix<F>{1,matrixB.M()};
+        }
+      }();
+
+      for (int i = 0; i < result.N(); ++i)
+        for (int j = 0; j < result.M(); ++j)
+          result[i][j] = matrixA.diagonal(i) * matrixB[i][j];
+      return result;
     }
 
   };
@@ -1107,8 +1166,11 @@ namespace Dune {
   struct DenseMatrixAssigner<DenseMatrix, DiagonalMatrix<field, N>> {
     static void apply(DenseMatrix& denseMatrix,
                       DiagonalMatrix<field, N> const& rhs) {
-      DUNE_ASSERT_BOUNDS(denseMatrix.M() == N);
-      DUNE_ASSERT_BOUNDS(denseMatrix.N() == N);
+#ifdef DUNE_CHECK_BOUNDS
+      if (denseMatrix.M() != N || denseMatrix.N() != N)
+        DUNE_THROW(Dune::RangeError, "Incompatible matrix dimensions in assignment.");
+#endif // DUNE_CHECK_BOUNDS
+
       denseMatrix = field(0);
       for (int i = 0; i < N; ++i)
         denseMatrix[i][i] = rhs.diagonal()[i];